webapp/backend/requirements.txt
autopep8
certifi
chardet
cycler
idna
jsonpatch
jsonpointer
kiwisolver
Pillow
pycodestyle
pyparsing
python-dateutil
pyzmq
requests
scipy
six
tqdm
urllib3
websocket-client
fastapi
uvicorn
torch
torchvision
torchaudio
numpy
pillow
matplotlib
opencv-python-headless